Мелкие доработки

+++ internal/pkg/auth/repository/auth_postgres.go
@@ -11,7 +11,7 @@ import (
 )
 
 // SetNewPasswordHash устанавливает пользователю новый хеш пароля
-func (r *AuthRepository) SetNewPasswordHash(ctx context.Context, userID int, newPasswordHash string) error {
+func (r *AuthRepository) SetNewPasswordHash(ctx context.Context, userID int64, newPasswordHash string) error {
 	query := `
 	UPDATE "user"
 	SET password_hash=$1
@@ -29,7 +29,7 @@ func (r *AuthRepository) SetNewPasswordHash(ctx context.Context, userID int, new
 }
 
 // GetUserPasswordHash получает хеш пароля пользователя
-func (r *AuthRepository) GetUserPasswordHash(ctx context.Context, userID int) (passwordHash *string, err error) {
+func (r *AuthRepository) GetUserPasswordHash(ctx context.Context, userID int64) (passwordHash *string, err error) {
 	query := `
 	SELECT password_hash
 	FROM "user"
